Refactored so only one detailed page is used for both IDIR and BCeID users. Fixed some issues where changes requested status goes directly to the edit screen even for IDIR users (#309)
